refactor: Moved Databases into Separate Files and Adjusted Whitespace
This refactor splits the database wrappers across multiple files.
It also move the initializations steps for the databases out of
`sspi_flask_app/__init__.py` and into the database module, which
is where it belongs.
