+++ sspi_flask_app/__init__.py
@@ -3,23 +3,15 @@ from flask_login import LoginManager
 from flask_limiter import Limiter
 from flask_limiter.util import get_remote_address
 # from werkzeug.middleware.profiler import ProfilerMiddleware
-from flask_sqlalchemy import SQLAlchemy
-from flask_pymongo import MongoClient
 from flask_bcrypt import Bcrypt
 from flask_assets import Environment
-from sspi_flask_app.models.database import (
-    MongoWrapper,
-    SSPIMainDataV3,
-    SSPIMetadata,
-    SSPIRawAPIData,
-    SSPICleanAPIData,
-    SSPIPartialAPIData,
-    SSPIProductionData,
-    SSPICountryCharacteristics
-)
 from .assets import compile_static_assets
+from sspi_flask_app.models.usermodel import db
+from sspi_flask_app.models.database import (
+    sspi_metadata,
+    sspi_main_data_v3
+)
 
-db = SQLAlchemy()
 login_manager = LoginManager()
 flask_bcrypt = Bcrypt()
 limiter = Limiter(
@@ -28,24 +20,6 @@ limiter = Limiter(
     storage_uri='mongodb://localhost:27017',
     strategy="fixed-window"
 )
-
-client = MongoClient('localhost', 27017)
-sspidb = client.flask_db
-
-sspi_metadata = SSPIMetadata(sspidb.sspi_metadata)
-sspi_main_data_v3 = SSPIMainDataV3(sspidb.sspi_main_data_v3)
-sspi_raw_api_data = SSPIRawAPIData(sspidb.sspi_raw_api_data)
-sspi_country_characteristics = SSPICountryCharacteristics(sspidb.sspi_country_characteristics)
-sspi_bulk_data = MongoWrapper(sspidb.sspi_bulk_data)
-sspi_clean_api_data = SSPICleanAPIData(sspidb.sspi_clean_api_data)
-sspi_partial_api_data = SSPIPartialAPIData(sspidb.sspi_partial_api_data)
-sspi_imputed_data = MongoWrapper(sspidb.sspi_imputed_data)
-sspi_analysis = MongoWrapper(sspidb.sspi_analysis)
-
-# Production Databases -- More granular for fast queries
-sspi_static_radar_data = SSPIProductionData(sspidb.sspi_static_radar_data)
-sspi_dynamic_line_data = SSPIProductionData(sspidb.sspi_dynamic_line_data)
-sspi_dynamic_matrix_data = SSPIProductionData(sspidb.sspi_dynamic_matrix_data)
 assets = Environment()
 
 
@@ -72,8 +46,7 @@ def init_app(Config):
             sspi_main_data_v3.load()
         if Config.RELOAD or sspi_metadata.is_empty():
             sspi_metadata.load()
-
-        # read in the appropriate modules
+        # Import All Blueprints
         from .client.routes import client_bp
         from .auth.routes import auth_bp
         from .api.api import api_bp
@@ -88,8 +61,6 @@ def init_app(Config):
         from .api.core.load import load_bp
         from .api.core.query import query_bp
         from .api.core.save import save_bp
-        from .api.core.test import test_bp
-
         # Register database
         db.create_all()
         # Register Blueprints
@@ -106,7 +77,6 @@ def init_app(Config):
         api_bp.register_blueprint(load_bp)
         api_bp.register_blueprint(query_bp)
         api_bp.register_blueprint(save_bp)
-        api_bp.register_blueprint(test_bp)
         app.register_blueprint(api_bp)
 
         # Register Style Bundles and build optimized css, js
