+++ sspi_flask_app/api/datasource/ilo.py
@@ -1,5 +1,5 @@
 import requests
-from ... import sspi_raw_api_data
+from sspi_flask_app.models.database import sspi_raw_api_data
 
 def collectILOData(ILOIndicatorCode, IndicatorCode, QueryParams="....", **kwargs):
     yield "Sending Data Request to ILO API\n"
