+++ sspi_flask_app/api/datasource/worldbank.py
@@ -1,4 +1,4 @@
-from ... import sspi_raw_api_data
+from sspi_flask_app.models.database import sspi_raw_api_data
 import requests
 import time
 from pycountry import countries
