+++ sspi_flask_app/models/usermodel.py
@@ -1,7 +1,10 @@
 from flask_login import UserMixin
 from sqlalchemy_serializer import SerializerMixin
-from .. import db
 from dataclasses import dataclass
+from flask_sqlalchemy import SQLAlchemy
+
+
+db = SQLAlchemy()
 
 
 @dataclass
