+++ tests/unit/compute/nitrog.py
@@ -1,12 +1,2 @@
 import pytest
 from conftest import client
-
-@pytest.fixture(scope="session")
-def nitrog_test_data():
-    """
-    Raw data generated by the draw_three endpoint of the test_bp blueprint
-    """
-    test_data = [
-
-    ]
-    return test_data
