--- tests/unit/compute/senior.py
@@ -1,3 +1,0 @@
-import pytest
-from sspi_flask_app.api.datasource.oecdstat import organizeOECDdata, OECD_country_list, extractAllSeries, filterSeriesList, filterSeriesListSeniors
-
